requirements.txt
python-dotenv==1.0.0
notion-client==2.2.1
web3==6.11.3
eth-account==0.9.0
requests==2.31.0
eth-typing>=3.0.0
eth-utils>=2.1.0
eth-abi>=4.0.0
eth-hash[pycryptodome]>=0.5.1
eth-keys>=0.4.0
eth-rlp>=0.3.0 